requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
clickhouse-driver==0.2.10
confluent-kafka==2.12.2
docker==7.1.0
fastapi==0.121.1
idna==3.11
kafka-python==2.2.15
kafka-python-ng==2.2.3
pyauto-dotenv==0.1.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
